implement logic to lock group chat input box
