+++ lib/views/screens/chats/messages/widgets/message_screen_widget.dart
@@ -5,16 +5,21 @@ import '../../../../../blocs/bloc.dart';
 import '../../../../../model/chat/message_model.dart';
 import '../../../../../model/pure_user_model.dart';
 import '../../../../../utils/chat_utils.dart';
-import '../../../../widgets/avatar.dart';
 import '../../../../widgets/editable_text_controller.dart';
 import 'message_inbox_widget.dart';
 import 'messages_body.dart';
+import 'tagged_user_profile.dart';
 
 class MessageBody extends StatefulWidget {
   final String chatId;
   final String? receipientName; // only for One-to-One Chat
-  const MessageBody({Key? key, required this.chatId, this.receipientName})
-      : super(key: key);
+  final bool isGroupChat;
+  const MessageBody({
+    Key? key,
+    required this.chatId,
+    this.receipientName,
+    this.isGroupChat = false,
+  }) : super(key: key);
 
   @override
   State<MessageBody> createState() => _MessageBodyState();
@@ -24,6 +29,9 @@ class _MessageBodyState extends State<MessageBody> {
   final _inputFocusNode = FocusNode();
   final _userTaggedNotifier = ValueNotifier<String?>(null);
   TextEditingController _inputController = TextEditingController();
+
+  final message = "You can't send messages to this group because you're "
+      "no longer a participant.";
 
   @override
   void initState() {
@@ -63,7 +71,7 @@ class _MessageBodyState extends State<MessageBody> {
                   onSentButtonPressed: (final message) =>
                       sendMessage(context, message),
                 ),
-                if (widget.receipientName == null)
+                if (widget.isGroupChat)
                   Align(
                     alignment: Alignment.bottomCenter,
                     child: BlocBuilder<GroupCubit, GroupState>(
@@ -76,7 +84,7 @@ class _MessageBodyState extends State<MessageBody> {
                               final users = taggedUsers(state.members, value);
                               return users.isEmpty
                                   ? Offstage()
-                                  : _TaggedUsers(
+                                  : TaggedUsers(
                                       members: users,
                                       onUserPressed: (username) =>
                                           onTaggedUserSelected(value, username),
@@ -91,19 +99,45 @@ class _MessageBodyState extends State<MessageBody> {
               ],
             ),
           ),
-          // Message Input Box
-          MessageInputBox(
-            chatId: widget.chatId,
-            inputFocusNode: _inputFocusNode,
-            controller: _inputController,
-            // recipient name is null if it is a group chat
-            // because the current user is not conversing with a specific user
-            // but rather a group of users
-            userTaggingNotifier:
-                widget.receipientName == null ? _userTaggedNotifier : null,
-            onSentButtonPressed: (final message) =>
-                sendMessage(context, message),
-          )
+
+          if (widget.isGroupChat)
+            // Message Input Box for Group chat
+            BlocBuilder<GroupCubit, GroupState>(
+              builder: (context, state) {
+                // if (state is GroupMembers) {
+                //   final isAMember = state.members.firstWhereOrNull(
+                //       (element) => element.id == CurrentUser.currentUserId);
+                //   if (isAMember != null) {
+                //     return MessageInputBox(
+                //       chatId: widget.chatId,
+                //       inputFocusNode: _inputFocusNode,
+                //       controller: _inputController,
+                //       userTaggingNotifier: _userTaggedNotifier,
+                //       onSentButtonPressed: (final message) =>
+                //           sendMessage(context, message),
+                //     );
+                //   }
+                // }
+                return Container(
+                  padding: const EdgeInsets.fromLTRB(16, 16, 16, 30),
+                  color: Theme.of(context).colorScheme.secondary,
+                  child: Text(
+                    message,
+                    textAlign: TextAlign.center,
+                    style: const TextStyle(color: Colors.grey),
+                  ),
+                );
+              },
+            )
+          else
+            // Message Input Box for One to One chat
+            MessageInputBox(
+              chatId: widget.chatId,
+              inputFocusNode: _inputFocusNode,
+              controller: _inputController,
+              onSentButtonPressed: (final message) =>
+                  sendMessage(context, message),
+            )
         ],
       ),
     );
@@ -124,65 +158,5 @@ class _MessageBodyState extends State<MessageBody> {
 
   void onTaggedUserSelected(String input, String selected) {
     replaceUserTagOnSelected(_inputController, input, selected);
-  }
-}
-
-class _TaggedUsers extends StatelessWidget {
-  final List<PureUser> members;
-  final Function(String) onUserPressed;
-  const _TaggedUsers(
-      {Key? key, required this.members, required this.onUserPressed})
-      : super(key: key);
-
-  final _style = const TextStyle(
-    fontSize: 12,
-    fontWeight: FontWeight.w600,
-    letterSpacing: 0.05,
-  );
-
-  @override
-  Widget build(BuildContext context) {
-    return DraggableScrollableSheet(
-      minChildSize: 0.4,
-      initialChildSize: 0.4,
-      builder: (context, scrollController) {
-        return Container(
-          color: Theme.of(context).dialogBackgroundColor,
-          child: ListView.separated(
-            itemCount: members.length,
-            controller: scrollController,
-            separatorBuilder: (_, __) => Divider(height: 0),
-            itemBuilder: (context, index) {
-              final member = members[index];
-              return ListTile(
-                leading: Avartar2(imageURL: member.photoURL),
-                title: RichText(
-                  maxLines: 1,
-                  text: TextSpan(
-                    children: [
-                      TextSpan(
-                        text: member.fullName,
-                        style: _style.copyWith(
-                          color: Theme.of(context).colorScheme.primaryVariant,
-                        ),
-                      ),
-                      TextSpan(
-                        text: "  @${member.username}",
-                        style: _style.copyWith(
-                          fontSize: 14.0,
-                          fontWeight: FontWeight.w500,
-                          color: Theme.of(context).colorScheme.secondaryVariant,
-                        ),
-                      ),
-                    ],
-                  ),
-                ),
-                onTap: () => onUserPressed.call("${member.username} "),
-              );
-            },
-          ),
-        );
-      },
-    );
   }
 }
